fix a cli problem
if i use the cli "spider" without any parameter, the cli.js would throw an error says "opts.files is undefined". I think if I do this, the cli should return a help message, so i add a if statement to find whether opts.files is undefined, if it is, console a help message and exit.    
I think this is a better way than throw an error directly :-)    
PS.spider-lang is cool and really want submit more pr : p